+++ Tests/Functional/Service/RegistrationManagerTest.php
@@ -140,7 +140,7 @@ final class RegistrationManagerTest extends FunctionalTestCase
 
         $this->subject->notifyOrganizers($registration);
 
-        $expectedExpression = '/' . $this->getLanguageService()->getLL('label_vacancies') . ': 1\\n*$/';
+        $expectedExpression = '/' . $this->translate('label_vacancies') . ': 1\\n*$/';
         self::assertRegExp($expectedExpression, $this->getTextBodyOfEmail($this->email));
     }